Update namespace in test project to match main project namespace
